app/requirements.txt
Flask
google-cloud-speech
google-cloud-storage
mysql-connector-python
gunicorn